Remove IFluidHTMLView from WebFlow
