+++ examples/data-objects/webflow/package.json
@@ -69,6 +69,7 @@
     "temp-directory": "nyc/.nyc_output"
   },
   "dependencies": {
+    "@fluid-example/example-utils": "^0.55.0",
     "@fluid-example/flow-util-lib": "^0.55.0",
     "@fluidframework/aqueduct": "^0.55.0",
     "@fluidframework/common-definitions": "^0.20.1",
